Add another rule for m4gic bots
There are a lot of bots going just by "m4gic" now.
Unrelated: There is also a weird unicode character that a lot of bots are using in their name; if I spot it again (haven't seen it in a couple days), I'll try to add it to the PR if I see it. I think names with this character should only be marked as suspicious as it may be used in non-bot usernames.